requirements.txt
pymysql
Flask_SQLAlchemy==3.0.3
Flask==2.2.3
PyYAML
